+++ frontend/src/api/index.ts
@@ -1,12 +1,15 @@
 export { default as axiosInstance } from '@api/axiosInstance';
 export { default as deleteReview } from '@api/deleteReview';
+export { default as getAccessToken } from '@api/getAccessToken';
 export { default as getMyStudyList } from '@api/getMyStudyList';
 export { default as getStudyDetail } from '@api/getStudyDetail';
 export { default as getStudyList } from '@api/getStudyList';
 export { default as getStudyReviews } from '@api/getStudyReviews';
 export { default as getTagList } from '@api/getTagList';
-export { default as putReview } from '@api/putReview';
-export { default as postAccessToken } from '@api/postAccessToken';
+export { default as getUserInformation } from '@api/getUserInformation';
+export { default as getUserRole } from '@api/getUserRole';
 export { default as postJoiningStudy } from '@api/postJoiningStudy';
+export { default as postLogin } from '@api/postLogin';
 export { default as postNewStudy } from '@api/postNewStudy';
 export { default as postReview } from '@api/postReview';
+export { default as putReview } from '@api/putReview';
